LayoutTests/platform/mac/fast/forms/select/select-disabled-expected.txt

Tests that user cannot change disabled select by keyboard

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select1.value is "a"
PASS select1.value is "b"
PASS select1.value is "b"
PASS select1.value is "b"
PASS select2.value is "a"
PASS select2.value is "b"
PASS select2.value is "b"
PASS select2.value is "b"
Tests on select disable, popup menu is hidden
FAIL internals.isSelectPopupVisible(select3) should be true. Was false.
PASS internals.isSelectPopupVisible(select3) is false
PASS internals.isSelectPopupVisible(select3) is false
Tests on select disable, value cannot be changed
PASS select4.value is "b"
PASS select4.value is "b"
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
